--- lldb/test/API/riscv/conflicting-extensions-disassembly/main.c
@@ -1,8 +1,0 @@
-extern void function_with_zcd_instructions();
-extern void function_with_zcmp_extension();
-
-int main() {
-  function_with_zcd_instructions();
-  function_with_zcmp_extension();
-  return 0;
-}
