--- lldb/test/API/riscv/disassembler/Makefile
@@ -1,8 +1,0 @@
-CXX_SOURCES := main.cpp
-
-all: a.out stripped.out
-
-include Makefile.rules
-
-stripped.out: a.out
-	$(OBJCOPY) --remove-section=.riscv.attributes a.out stripped.out
