+++ lldb/test/API/riscv/disassembler/TestDisassembler.py
@@ -6,12 +6,12 @@ import lldb
 from lldbsuite.test.decorators import *
 from lldbsuite.test.lldbtest import *
 from lldbsuite.test import lldbutil
+import os
 
 
 class TestDisassembler(TestBase):
     expected_zbb_instrs = ["andn", "orn", "xnor", "rol", "ror"]
 
-    @skipIfLLVMTargetMissing("RISCV")
     def test_without_riscv_attributes(self):
         """
         Tests disassembly of a riscv binary without the .riscv.attributes.
@@ -19,9 +19,11 @@ class TestDisassembler(TestBase):
         handle the bitmanip extension, so it is not expected to see zbb instructions
         in the output.
         """
-        self.build(dictionary={"CFLAGS_EXTRAS": "-march=rv64gc_zbb"})
+        yaml = os.path.join(self.getSourceDir(), "stripped.out.yaml")
+        exe = self.getBuildArtifact("stripped.out")
+        self.yaml2obj(yaml, exe)
 
-        target = self.dbg.CreateTarget(self.getBuildArtifact("stripped.out"))
+        target = self.dbg.CreateTarget(exe)
 
         self.expect("disassemble --name do_zbb_stuff")
         output = self.res.GetOutput()
@@ -37,12 +39,13 @@ class TestDisassembler(TestBase):
             "Instructions from the Zbb extension should be displayed as <unknown>",
         )
 
-    @skipIfLLVMTargetMissing("RISCV")
     def test_with_riscv_attributes(self):
         """
         Tests disassembly of a riscv binary with the .riscv.attributes.
         """
-        self.build(dictionary={"CFLAGS_EXTRAS": "-march=rv64gc_zbb"})
+        yaml = os.path.join(self.getSourceDir(), "a.out.yaml")
+        exe = self.getBuildArtifact("a.out")
+        self.yaml2obj(yaml, exe)
 
         target = self.dbg.CreateTarget(self.getBuildArtifact("a.out"))
 
@@ -51,3 +54,26 @@ class TestDisassembler(TestBase):
 
         for instr in self.expected_zbb_instrs:
             self.assertTrue(instr in output, "Invalid disassembler output")
+
+    def test_conflicting_extensions(self):
+        """
+        This test demonstrates the scenario where:
+        1. file_with_zcd.c is compiled with rv64gc (includes C and D).
+        2. file_with_zcmp.c is compiled with rv64imad_zcmp (includes Zcmp).
+        3. The linker merges .riscv.attributes, creating the union: C + D + Zcmp.
+
+        The Zcmp extension is incompatible with the C extension when the D extension is enabled.
+        Therefore, the arch string contains conflicting extensions, and LLDB should
+        display an appropriate warning in this case.
+        """
+        yaml = os.path.join(self.getSourceDir(), "conflicting.out.yaml")
+        exe = self.getBuildArtifact("conflicting.out")
+        self.yaml2obj(yaml, exe)
+
+        target = self.dbg.CreateTarget(self.getBuildArtifact("a.out"))
+        output = self.res.GetOutput()
+
+        self.assertIn(
+            output,
+            "The .riscv.attributes section contains an invalid RISC-V arch string",
+        )
