+++ lldb/test/API/riscv/disassembler/a.out.yaml
@@ -1,0 +1,32 @@
+--- !ELF
+FileHeader:
+  Class:           ELFCLASS64
+  Data:            ELFDATA2LSB
+  Type:            ET_DYN
+  Machine:         EM_RISCV
+  Flags:           [ EF_RISCV_RVC, EF_RISCV_FLOAT_ABI_DOUBLE ]
+Sections:
+  - Name:            .text
+    Type:            SHT_PROGBITS
+    Flags:           [ SHF_ALLOC, SHF_EXECINSTR ]
+    Address:         0x550
+    AddressAlign:    0x4
+    Content:         EF002002AA87172500000335A5A782653000137101FF814601470A88EFF05FFD029097210000938161298280010017250000130525A9972700009387A7A86388A7009727000083B7E7A491C38287828017250000130505A797250000938585A6898D93D73540FD91BE95858599C59727000083B7A7A291C3828782809727000083C747A49DE7411106E49727000083B7E79F91C717250000033545A28297EFF01FF9A260854717270000230DF7A041018280828071BF411106E422E000083376B5403366B5403346B5403316B5603356B560A260026441018280011106EC22E8001001452330A4FE2326A4FEEFF0BFFC033504FEE260426405618280
+  - Name:            .riscv.attributes
+    Type:            SHT_RISCV_ATTRIBUTES
+    AddressAlign:    0x1
+    Content:         416C000000726973637600016200000004100572763634693270315F6D3270305F613270315F663270325F643270325F633270305F7A696373723270305F7A6966656E6365693270305F7A6D6D756C3170305F7A61616D6F3170305F7A616C7273633170305F7A626231703000
+Symbols:
+  - Name:            _Z12do_zbb_stuffv
+    Type:            STT_FUNC
+    Section:         .text
+    Binding:         STB_GLOBAL
+    Value:           0x606
+    Size:            0x24
+  - Name:            main
+    Type:            STT_FUNC
+    Section:         .text
+    Binding:         STB_GLOBAL
+    Value:           0x62A
+    Size:            0x22
+...
